+++ Source/EventFlow.AspNetCore.Tests/IntegrationTests/Site/SiteTests.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015-2018 Rasmus Mikkelsen
-// Copyright (c) 2015-2018 eBay Software Foundation
+// Copyright (c) 2015-2019 Rasmus Mikkelsen
+// Copyright (c) 2015-2019 eBay Software Foundation
 // https://github.com/eventflow/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -22,15 +22,22 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net.Http;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
+using EventFlow.Configuration.Serialization;
+using EventFlow.EventStores;
 using EventFlow.Logs;
 using EventFlow.TestHelpers;
+using EventFlow.TestHelpers.Aggregates;
 using EventFlow.TestHelpers.Aggregates.Commands;
 using FluentAssertions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
+using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
 using NUnit.Framework;
 
@@ -39,17 +46,21 @@ namespace EventFlow.AspNetCore.Tests.IntegrationTests.Site
 	[Category(Categories.Integration)]
 	public class SiteTests : Test
 	{
-		private TestServer _server;
 		private HttpClient _client;
+        private IEventStore _eventStore;
+        private IJsonOptions _jsonOptions;
 		private ConsoleLog _log;
+        private TestServer _server;
 
-		[SetUp]
+	    [SetUp]
 		public void SetUp()
 		{
 			_server = new TestServer(new WebHostBuilder()
 				.UseStartup<Startup>());
 			_client = _server.CreateClient();
 			_log = new ConsoleLog();
+            _jsonOptions = GetService<IJsonOptions>();
+            _eventStore = GetService<IEventStore>();
 		}
 
 		[TearDown]
@@ -65,6 +76,23 @@ namespace EventFlow.AspNetCore.Tests.IntegrationTests.Site
 			await GetAsync("thingy/ping?id=thingy-d15b1562-11f2-4645-8b1a-f8b946b566d3").ConfigureAwait(false);
 			await GetAsync("thingy/ping?id=thingy-d15b1562-11f2-4645-8b1a-f8b946b566d3").ConfigureAwait(false);
 		}
+
+	    [Test]
+	    public async Task ValidSingleValue()
+	    {
+            // Act
+	        var result = await GetAsync("thingy/singlevalue/123").ConfigureAwait(false);
+	        result.Should().Be("123");
+	    }
+
+	    [Test]
+	    public void InvalidSingleValue()
+	    {
+	        // Arrange + Act
+	        Func<Task> call = async () => await GetAsync("thingy/singlevalue/asdf").ConfigureAwait(false);
+
+	        call.Should().Throw<HttpRequestException>();
+	    }
 
 		[Test]
 		public async Task PublishCommand()
@@ -84,6 +112,33 @@ namespace EventFlow.AspNetCore.Tests.IntegrationTests.Site
 
 			action.Should().Throw<HttpRequestException>("because of command is null.");
 		}
+
+        [Test]
+        public async Task EventsContainMetadata()
+        {
+            var id = ThingyId.New;
+            // Arrange + Act
+            await GetAsync($"thingy/ping?id={id}").ConfigureAwait(false);
+
+            // Assert
+            var events = await _eventStore.LoadEventsAsync<ThingyAggregate, ThingyId>(id, CancellationToken.None);
+            var metadata = events.Single().Metadata;
+
+            metadata.Should().Contain(new[]
+                {
+                    new KeyValuePair<string, string>("user_claim[sid]", "test-sid"),
+                    new KeyValuePair<string, string>("user_claim[role]", "test-role-1;test-role-2")
+                }
+            );
+
+            metadata.Should().NotContain(
+                new KeyValuePair<string, string>("user_claim[name]", "test-name"));
+        }
+
+        private T GetService<T>()
+        {
+            return _server.Host.Services.GetRequiredService<T>();
+        }
 
 		private async Task<string> GetAsync(string url)
 		{
@@ -106,7 +161,9 @@ namespace EventFlow.AspNetCore.Tests.IntegrationTests.Site
 
 		private async Task<string> PostAsync(string url, object obj)
 		{
-			var json = JsonConvert.SerializeObject(obj);
+		    var settings = new JsonSerializerSettings();
+            _jsonOptions.Apply(settings);
+			var json = JsonConvert.SerializeObject(obj, settings);
 			var stringContent = new StringContent(json, Encoding.UTF8, "application/json");
 		
 			var response = await _client.PostAsync(url, stringContent);