+++ RELEASE_NOTES.md
@@ -8,6 +8,31 @@
       .AddConverter<SomeConverter>()
     )
   ```
+* New: ASP.NET Core enhancements:
+  - New fluent configuration API for ASP.NET Core components:
+    `services.AddEventFlow(o => o.AddAspNetCore(c => {...}));` (old syntax
+    `AddAspNetCoreMetadataProviders` is now deprecated).
+  - `.RunBootstrapperOnHostStartup()` runs bootstrappers together with ASP.NET
+    host startup. Previously, this was done in `AddAspNetCoreMetadataProviders`
+    and led to some confusion.
+  - `.UseMvcJsonOptions()` adds EventFlow JSON configuration (see below) to ASP.NET Core,
+    so you can accept and return Single Value Objects as plain strings for example.
+  - `.Add{Whatever}Metadata()` configures specific metadata provider.
+  - `.AddUserClaimsMetadata(params string claimTypes)` configures the new claims metadata
+    provider (for auditing or "ChangedBy" in read models).
+  - `.UseLogging()` configures an adapter for Microsoft.Extensions.Logging
+  - `.UseModelBinding()` adds model binding support for Single Value Objects:
+    ```csharp
+	    [HttpGet("customers/{id}")]
+	    public async Task<IActionResult> SingleValue(CustomerId id)
+	    {
+	        if (!ModelState.IsValid)
+	        {
+	            return BadRequest(ModelState);
+	        }
+    ```
+* Fix: ASP.NET Core `AddRequestHeadersMetadataProvider` doesn't throw when
+  HttpContext is null.
 
 ### New in 0.72.3914 (released 2019-05-28)